+++ test/validator/body.test.ts
@@ -416,7 +416,7 @@ describe('Body Validator', () => {
 		expect(value).toBe('number')
 	})
 
-	it("don't coerce number to numeric", async () => {
+	it("coerce number to numeric", async () => {
 		const app = new Elysia().post('/', ({ body }) => typeof body, {
 			body: t.Number()
 		})
@@ -431,7 +431,7 @@ describe('Body Validator', () => {
 			})
 		)
 
-		expect(response.status).toBe(422)
+		expect(response.status).toBe(200)
 	})
 
 	it("don't coerce number object to numeric", async () => {