+++ src/index.ts
@@ -19,6 +19,7 @@ import {
 
 import {
 	cloneInference,
+	coercePrimitiveRoot,
 	deduplicateChecksum,
 	fnToContainer,
 	localHookToLifeCycleStore,
@@ -507,7 +508,8 @@ export default class Elysia<
 						body: getSchemaValidator(cloned.body, {
 							dynamic,
 							models,
-							normalize
+							normalize,
+							additionalCoerce: coercePrimitiveRoot()
 						}),
 						headers: getSchemaValidator(cloned.headers, {
 							dynamic,
@@ -545,7 +547,8 @@ export default class Elysia<
 								{
 									dynamic,
 									models,
-									normalize
+									normalize,
+									additionalCoerce: coercePrimitiveRoot()
 								}
 							))
 						},