+++ src/utils.ts
@@ -267,8 +267,9 @@ export const mergeHook = (
 
 interface ReplaceSchemaTypeOptions {
 	from: TSchema
-	to(): TSchema
+	to(options: Object): TSchema
 	excludeRoot?: boolean
+	rootOnly?: boolean
 	/**
 	 * Traverse until object is found except root object
 	 **/
@@ -333,7 +334,7 @@ const _replaceSchemaType = (
 	if (schema[Kind] === fromSymbol) {
 		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		const { anyOf, oneOf, allOf, not, properties, items, ...rest } = schema
-		const to = options.to()
+		const to = options.to(rest)
 
 		// If t.Transform is used, we need to re-calculate Encode, Decode
 		let transform
@@ -472,13 +473,13 @@ const _replaceSchemaType = (
 
 	const properties = schema?.properties as Record<string, TSchema>
 
-	if (properties)
+	if (properties && root && options.rootOnly !== true)
 		for (const [key, value] of Object.entries(properties)) {
 			switch (value[Kind]) {
 				case fromSymbol:
 					// eslint-disable-next-line @typescript-eslint/no-unused-vars
 					const { anyOf, oneOf, allOf, not, type, ...rest } = value
-					const to = options.to()
+					const to = options.to(rest)
 
 					if (to.anyOf)
 						for (let i = 0; i < to.anyOf.length; i++)
@@ -554,22 +555,31 @@ export const getSchemaValidator = <T extends TSchema | string | undefined>(
 
 	let schema: TSchema = typeof s === 'string' ? models[s] : s
 
-	if (coerce)
-		schema = replaceSchemaType(schema, [
-			{
-				from: t.Number(),
-				to: () => t.Numeric(),
-				untilObjectFound: true
-			},
-			{
-				from: t.Boolean(),
-				to: () => t.BooleanString(),
-				untilObjectFound: true
-			},
-			...(Array.isArray(additionalCoerce)
-				? additionalCoerce
-				: [additionalCoerce])
-		])
+	if (coerce || additionalCoerce) {
+		if (coerce)
+			schema = replaceSchemaType(schema, [
+				{
+					from: t.Number(),
+					to: (options) => t.Numeric(options),
+					untilObjectFound: true
+				},
+				{
+					from: t.Boolean(),
+					to: (options) => t.BooleanString(options),
+					untilObjectFound: true
+				},
+				...(Array.isArray(additionalCoerce)
+					? additionalCoerce
+					: [additionalCoerce])
+			])
+		else {
+			schema = replaceSchemaType(schema, [
+				...(Array.isArray(additionalCoerce)
+					? additionalCoerce
+					: [additionalCoerce])
+			])
+		}
+	}
 
 	// console.dir(schema, {
 	// 	depth: null
@@ -810,7 +820,28 @@ export const stringToStructureCoercions = () => {
 			}
 		] satisfies ReplaceSchemaTypeOptions[]
 	}
+
 	return _stringToStructureCoercions
+}
+
+let _coercePrimitiveRoot: ReplaceSchemaTypeOptions[]
+
+export const coercePrimitiveRoot = () => {
+	if (!_coercePrimitiveRoot)
+		_coercePrimitiveRoot = [
+			{
+				from: t.Number(),
+				to: (options) => t.Numeric(options),
+				rootOnly: true
+			},
+			{
+				from: t.Boolean(),
+				to: (options) => t.BooleanString(options),
+				rootOnly: true
+			}
+		] satisfies ReplaceSchemaTypeOptions[]
+
+	return _coercePrimitiveRoot
 }
 
 export const getCookieValidator = ({
@@ -1469,48 +1500,47 @@ type PropertyKeys<T> = {
 
 type PropertiesOnly<T> = Pick<T, PropertyKeys<T>>
 
-export const classToObject = <T>(
-	instance: T,
-	processed: WeakMap<object, object> = new WeakMap()
-): T extends object ? PropertiesOnly<T> : T => {
-	if (typeof instance !== 'object' || instance === null)
-		return instance as any
+// export const classToObject = <T>(
+// 	instance: T,
+// 	processed: WeakMap<object, object> = new WeakMap()
+// ): T extends object ? PropertiesOnly<T> : T => {
+// 	if (typeof instance !== 'object' || instance === null)
+// 		return instance as any
 
-	if (Array.isArray(instance))
-		return instance.map((x) => classToObject(x, processed)) as any
+// 	if (Array.isArray(instance))
+// 		return instance.map((x) => classToObject(x, processed)) as any
 
-	if (processed.has(instance)) return processed.get(instance) as any
+// 	if (processed.has(instance)) return processed.get(instance) as any
 
-	const result: Partial<T> = {}
+// 	const result: Partial<T> = {}
 
-	for (const key of Object.keys(instance) as Array<keyof T>) {
-		const value = instance[key]
-		if (typeof value === 'object' && value !== null)
-			result[key] = classToObject(value, processed) as T[keyof T]
-		else result[key] = value
-	}
+// 	for (const key of Object.keys(instance) as Array<keyof T>) {
+// 		const value = instance[key]
+// 		if (typeof value === 'object' && value !== null)
+// 			result[key] = classToObject(value, processed) as T[keyof T]
+// 		else result[key] = value
+// 	}
 
-	const prototype = Object.getPrototypeOf(instance)
-	if (!prototype) return result as any
+// 	const prototype = Object.getPrototypeOf(instance)
+// 	if (!prototype) return result as any
 
-	const properties = Object.getOwnPropertyNames(prototype)
+// 	const properties = Object.getOwnPropertyNames(prototype)
 
-	for (const property of properties) {
-		const descriptor = Object.getOwnPropertyDescriptor(
-			Object.getPrototypeOf(instance),
-			property
-		)
+// 	for (const property of properties) {
+// 		const descriptor = Object.getOwnPropertyDescriptor(
+// 			Object.getPrototypeOf(instance),
+// 			property
+// 		)
 
-		if (descriptor && typeof descriptor.get === 'function') {
-			// ? Very important to prevent prototype pollution
-			if (property === '__proto__') continue
+// 		if (descriptor && typeof descriptor.get === 'function') {
+// 			// ? Very important to prevent prototype pollution
+// 			if (property === '__proto__') continue
 
-			console.log(property)
-			;(result as any)[property as keyof typeof instance] = classToObject(
-				instance[property as keyof typeof instance]
-			)
-		}
-	}
+// 			;(result as any)[property as keyof typeof instance] = classToObject(
+// 				instance[property as keyof typeof instance]
+// 			)
+// 		}
+// 	}
 
-	return result as any
-}
+// 	return result as any
+// }